Put component test into folder
