+++ ui/tests/components/ResourceGroup.test.js
@@ -2,16 +2,16 @@ import React from "react";
 import Adapter from "enzyme-adapter-react-16";
 import { shallow, configure } from "enzyme";
 import Button from "@material-ui/core/Button";
-import SubpolicyGroup from "../src/components/SubpolicyGroup";
-import Subpolicy from "../src/containers/Subpolicy";
+import ResourceGroup from "../../src/components/ResourceGroup";
+import Resource from "../../src/containers/Resource";
 
 function shallowSetup() {
   const props = {
-    subpolicies: ["a", "b", "c"],
-    createSubpolicy: jest.fn()
+    resources: ["a", "b"],
+    createResource: jest.fn()
   };
 
-  const enzymeWrapper = shallow(<SubpolicyGroup {...props} />);
+  const enzymeWrapper = shallow(<ResourceGroup {...props} />);
   return {
     props,
     enzymeWrapper
@@ -19,22 +19,20 @@ function shallowSetup() {
 }
 
 configure({ adapter: new Adapter() });
-describe("SubpolicyGroup", () => {
+describe("ResourceGroup", () => {
   describe("renders", () => {
     it("it should render correct number of resources", () => {
       const { enzymeWrapper } = shallowSetup();
-      expect(enzymeWrapper.find(Subpolicy).length).toEqual(3);
+      expect(enzymeWrapper.find(Resource).length).toEqual(2);
     });
   });
 
   describe("clicks", () => {
     it("clicking delete should trigger createResource", () => {
       const { enzymeWrapper, props } = shallowSetup();
-      expect(enzymeWrapper.find(Button).length).toEqual(1);
-
       const button = enzymeWrapper.find(Button).first();
       button.simulate("click");
-      expect(props.createSubpolicy.mock.calls.length).toBe(1);
+      expect(props.createResource.mock.calls.length).toBe(1);
     });
   });
 });
